requirements.txt
transformers>=4.0.0
torch>=2.0.0
llama-cpp-python[server]==0.2.27
pyOpenSSL>=23.0.0
cryptography>=41.0.0
flask>=2.0.0
werkzeug>=2.0.0
flask-swagger-ui>=3.36.0
requests>=2.25.0
waitress>=2.1.0
# Remove: fastapi, uvicorn, pydantic, aiohttp
